Add support for excluding X11 headers.
This is enabled with GLBIND_NO_XLIB_HEADERS. When this is used, glbind
will declare it's own versions of the necessary types and identifiers.
For structures it's expected that you'll provide pointers to your own
versions which are compatible with the normal Xlib structures. You will
need to manually cast them to the glbind_* types where necessary.