Expose core GLX functions to global scope.
I think the reason I avoided this before was because earlier versions
of glbind included glx.h directly. But since that is no longer the case
I think it should be OK to expose these. If this creates issues later
we can probably wrap this in some kind of compile time option.